+++ entry.sh
@@ -30,44 +30,23 @@ image_type=$(cat /image-type)
 case "$image_type" in
     carbon-c-relay)
         assert_conf_exist relay.conf
-        exec su-exec openmetric carbon-c-relay -f $CONF/relay.conf -l $LOG/relay.log -c '-_:#@'
+        exec su-exec openmetric carbon-c-relay -f $CONF/relay.conf -c '-_:#@$'
         ;;
     go-carbon)
         assert_conf_exist carbon.conf
         assert_conf_exist schemas.conf
         exec su-exec openmetric go-carbon --config $CONF/carbon.conf
         ;;
-    carbonzipper)
-        assert_conf_exist zipper.conf
-
-        if carbonzipper -h 2>&1 | grep -q -- '-stdout'; then
-            # current stable release
-            exec su-exec openmetric carbonzipper -c $CONF/zipper.conf -logdir /openmetric/log/
-        else
-            # edge release
-            exec su-exec openmetric carbonzipper -c $CONF/zipper.conf
-        fi
-        ;;
     carbonapi)
-        assert_conf_exist api.conf
-
-        # check if current version support configuration file
-        # NOTE although we use 'api.conf' as filename for both version, file format is different.
-        # For current stable release, api.conf should be bash source-able.
-        # For edge release, api.conf is yaml.
-        if carbonapi -h 2>&1 | grep -q -- '-config'; then
-            # edge release, support conf file
-            exec su-exec openmetric carbonapi -config $CONF/api.conf
+        if test -f "$CONF/api.yaml"; then
+            echo "Using config file $CONF/api.yaml"
+            exec su-exec openmetric carbonapi -config $CONF/api.yaml
+        elif test -f "$CONF/api.toml"; then
+            echo "Using config file: $CONF/api.toml"
+            exec su-exec openmetric carbonapi -config $CONF/api.toml
         else
-            # current stable release
-            source $CONF/api.conf
-            exec su-exec openmetric carbonapi \
-                -p ${LISTEN_PORT:-5000} \
-                -z ${ZIPPER_URL:-127.0.0.1:8080} \
-                -graphite ${GRAPHITE_URL:-127.0.0.1:2003} \
-                -prefix ${INTERNAL_METRIC_PREFIX:-carbon.api} \
-                -i ${INTERVAL:-10s} \
-                -logdir ${LOGDIR:-/openmetric/log/}
+            echo "Config file not found, nither api.yaml or api.toml"
+            exit 1
         fi
         ;;
     grafana)